Migrate monetary amounts from float to Decimal
Bank statement totals were accumulated with floating-point arithmetic
across debit/credit sums and the reconcile test had already grown a
0.05 tolerance to accommodate drift. Decimal eliminates that class of
bug.

Changes:
- Transaction.amount is now Decimal; _parse_amount returns Decimal.
- API exposes a Money type (Annotated[Decimal, PlainSerializer]) that
  serializes as a fixed-decimal JSON string to prevent precision loss
  on the wire.
- Frontend types updated: amount and totals fields are 'string' now.
  formatMoney parses to Number only at display time; arithmetic
  callers (none yet) must do the same explicitly.
- Smoke test TOLERANCE drops to Decimal('0.00') for exact reconcile.
- Unit test assertions switch to Decimal('...') comparisons.

+++ packages/parsers/kasa_parsers/cimb_cc/parser.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import re
 from datetime import date
+from decimal import Decimal
 from pathlib import Path
 
 from ..core import (
@@ -223,5 +224,5 @@ def _parse_date(s: str, stmt_year: int, stmt_month: int) -> date:
     return date(year, month, day)
 
 
-def _parse_amount(s: str) -> float:
-    return float(s.replace(",", ""))
+def _parse_amount(s: str) -> Decimal:
+    return Decimal(s.replace(",", ""))
